+++ core/src/main/kotlin/dark/ecs/systems/FlashlightRayTracingSystem.kt
@@ -2,14 +2,13 @@ package dark.ecs.systems
 
 import com.badlogic.ashley.core.Entity
 import com.badlogic.ashley.systems.IteratingSystem
-import com.badlogic.gdx.math.Vector2
 import com.badlogic.gdx.physics.box2d.Fixture
 import com.badlogic.gdx.physics.box2d.World
 import dark.core.GameSettings
-import dark.ecs.components.blobcomponents.Blob
 import dark.ecs.components.Flashlight
 import dark.ecs.components.PropsAndStuff
-import eater.ecs.ashley.components.GeneralMemory
+import dark.ecs.components.blobcomponents.Blob
+import dark.ecs.systems.stackai.MemoryEvent
 import eater.ecs.ashley.components.Memory
 import eater.ecs.ashley.components.Remove
 import eater.ecs.ashley.components.TransformComponent
@@ -20,7 +19,6 @@ import ktx.ashley.allOf
 import ktx.ashley.exclude
 import ktx.box2d.RayCast
 import ktx.box2d.rayCast
-import ktx.log.info
 import ktx.math.plus
 import ktx.math.times
 import ktx.math.vec2
@@ -87,6 +85,3 @@ class FlashlightRayTracingSystem(private val world: World) : IteratingSystem(
     }
 }
 
-sealed class MemoryEvent : GeneralMemory {
-    class HitByLight(val lightSource: Vector2) : MemoryEvent()
-}
