+++ core/src/main/kotlin/dark/ecs/systems/stackai/StackAiSystem.kt
@@ -1,4 +1,4 @@
-package dark.ecs.systems
+package dark.ecs.systems.stackai
 
 import com.badlogic.ashley.core.Entity
 import com.badlogic.ashley.systems.IteratingSystem
